feature/img-filter Added some image unit tests
A lot of functions in the image service are harder to test due the
dependency of the Jimp library.

These function are using simply call to another function making the
unit test lose some of its value, since we will be testing the library
function and anything other important logic.

Some function were moved to the context layer to make flexible to
use the full potential of the Jimp library to adapt to different business.